+++ kernel/arch/dreamcast/include/arch/irq.h
@@ -26,6 +26,7 @@
 #ifndef __ARCH_IRQ_H
 #define __ARCH_IRQ_H
 
+#include <stdalign.h>
 #include <stdbool.h>
 #include <stdint.h>
 #include <sys/cdefs.h>
@@ -48,6 +49,15 @@ __BEGIN_DECLS
     interrupts are utilized by various KOS drivers and have higher-level APIs
     for hooking into them. Care must be taken to not interfere with the IRQ
     handling which is being done by in-use KOS drivers.
+
+    \note
+    The naming convention used by this API differs from that of the actual SH4
+    manual for historical reasons (it wasn't platform-specific). The SH4 manual
+    refers to the most general type of CPU events which result in a SW
+    callback as "exceptions," with "interrupts" and "general exceptions" being
+    subtypes of exceptions. This API uses the term "interrupt" and "exception"
+    interchangeably, except where it is explicitly noted that "SH4 interrupts"
+    or "SH4 general exceptions" are being referred to, more specifically.
 
     @{
 */
@@ -170,11 +180,11 @@ void irq_create_context(irq_context_t *context, uint32_t stack_pointer,
 
 /** Interrupt exception codes
 
-   Dreamcast-specific exception codes. Used to identify the source or type of
-   an interrupt. Each exception code is of a certain "type" which dictates how the interrupt
-   is generated and handled.
+   SH-specific exception codes. Used to identify the source or type of an
+   interrupt. Each exception code is of a certain "type" which dictates how the
+   interrupt is generated and handled.
 
-    List of exception types:
+   List of exception types:
 
    |Type    | Description
    |--------|------------
@@ -309,13 +319,28 @@ int irq_inside_int(void);
 
 /** Type representing an interrupt mask state. */
 typedef uint32_t irq_mask_t;
+
+/** Get status register contents.
+
+    Returns the current value of the status register, as irq_disable() does.
+    The function can be found in arch\dreamcast\kernel\entry.s
+
+    \note
+    This is the entire status register word, not just the `IMASK` field.
+
+    \retval                 Status register word
+    \sa irq_disable()
+*/
+irq_mask_t irq_get_sr(void);
+
 /** Disable interrupts.
 
-    This function will disable interrupts, but will leave exceptions enabled.
+    This function will disable SH4 interrupts, but will leave SH4 general
+    exceptions enabled.
 
-    \return                 The state of IRQs before calling the function. This
-                            can be used to restore this state later on with
-                            irq_restore().
+    \return                 The state of the SH4 interrupts before calling the
+                            function. This can be used to restore this state
+                            later on with irq_restore().
 
     \sa irq_restore(), irq_enable()
 */
@@ -340,6 +365,14 @@ void irq_enable(void);
     \sa irq_disable()
 */
 void irq_restore(irq_mask_t v);
+
+/** \brief  Disable interrupts with scope management.
+
+    This macro will disable interrupts, similarly to irq_disable(), with the
+    difference that the interrupt state will automatically be restored once the
+    execution exits the functional block in which the macro was called.
+*/
+#define irq_disable_scoped() __irq_disable_scoped(__LINE__)
 
 /** @} */
 
@@ -383,6 +416,16 @@ void irq_force_return(void);
 */
 typedef void (*irq_handler)(irq_t code, irq_context_t *context, void *data);
 
+
+/** The type of a full callback of an IRQ handler and userdata.
+
+    This type is used to set or get IRQ handlers and their data.
+*/
+typedef struct irq_cb {
+    irq_handler hdl;    /**< A pointer to a procedure to handle an exception. */
+    void       *data;   /**< A pointer that will be passed along to the callback. */
+} irq_cb_t;
+
 /** \defgroup irq_handlers_ind  Individual
     \brief                      API for managing individual IRQ handlers.
 
@@ -410,12 +453,13 @@ int irq_set_handler(irq_t code, irq_handler hnd, void *data);
 /** Get the address of the current handler for the IRQ type.
 
     \param  code            The IRQ type to look up.
-    
-    \return                 A pointer to the procedure to handle the exception.
+
+    \return                 The current handler for the IRQ type and
+                            its userdata.
 
     \sa irq_set_handler()
 */
-irq_handler irq_get_handler(irq_t code);
+irq_cb_t irq_get_handler(irq_t code);
 
 /** @} */
 
@@ -442,10 +486,10 @@ int irq_set_global_handler(irq_handler handler, void *data);
 
 /** Get the global exception handler.
 
-    \return                 The global exception handler set with
+    \return                 The global exception handler and userdata set with
                             irq_set_global_handler(), or NULL if none is set.
 */
-irq_handler irq_get_global_handler(void);
+irq_cb_t irq_get_global_handler(void);
 /** @} */
 
 /** @} */
@@ -477,14 +521,6 @@ static inline void __irq_scoped_cleanup(int *state) {
 
 #define __irq_disable_scoped(l) ___irq_disable_scoped(l)
 /** \endcond */
-
-/** \brief  Disable interrupts with scope management.
-
-    This macro will disable interrupts, similarly to irq_disable(), with the
-    difference that the interrupt state will automatically be restored once the
-    execution exits the functional block in which the macro was called.
-*/
-#define irq_disable_scoped() __irq_disable_scoped(__LINE__)
 
 /** \brief  Minimum/maximum values for IRQ priorities