+++ kernel/arch/dreamcast/kernel/irq.c
@@ -30,12 +30,6 @@
 /* Interrupt priority registers */
 #define REG_IPR(x) ( *((volatile uint16_t *)(0xffd00004 + (x) * 4)) )
 
-/* IRQ handler closure */
-struct irq_cb {
-    irq_handler hdl;
-    void       *data;
-};
-
 /* TRAPA handler closure */
 struct trapa_cb {
     trapa_handler hdl;
@@ -43,16 +37,16 @@ struct trapa_cb {
 };
 
 /* Individual exception handlers */
-static struct irq_cb   irq_handlers[0x40];
+static irq_cb_t        irq_handlers[0x40];
 /* TRAPA exception handlers */
 static struct trapa_cb trapa_handlers[0x100];
 
 /* Global exception handler -- hook this if you want to get each and every
    exception; you might get more than you bargained for, but it can be useful. */
-static struct irq_cb   global_irq_handler;
+static irq_cb_t        global_irq_handler;
 
 /* Default IRQ context location */
-static irq_context_t    irq_context_default;
+static irq_context_t   irq_context_default;
 
 /* Are we inside an interrupt? */
 static int inside_int;
@@ -73,14 +67,14 @@ int irq_set_handler(irq_t code, irq_handler hnd, void *data) {
 }
 
 /* Get the address of the current handler */
-irq_handler irq_get_handler(irq_t code) {
+irq_cb_t irq_get_handler(irq_t code) {
     /* Make sure they don't do something crackheaded */
     if(code >= 0x1000 || (code & 0x000f))
-        return NULL;
+        return (irq_cb_t){ NULL, NULL };
 
     code >>= 5;
 
-    return irq_handlers[code].hdl;
+    return irq_handlers[code];
 }
 
 /* Set a global handler */
@@ -91,8 +85,8 @@ int irq_set_global_handler(irq_handler hnd, void *data) {
 }
 
 /* Get the global exception handler */
-irq_handler irq_get_global_handler(void) {
-    return global_irq_handler.hdl;
+irq_cb_t irq_get_global_handler(void) {
+    return global_irq_handler;
 }
 
 /* Set or remove a trapa handler */
@@ -251,23 +245,6 @@ void irq_handle_exception(int code) {
         handled = 1;
     }
 
-    /* dbgio_printf("got int %04x %04x\n", code, evt); */
-
-    /* If it's a timer interrupt, clear the status */
-    if(evt >= EXC_TMU0_TUNI0 && evt <= EXC_TMU2_TUNI2) {
-        if(evt == EXC_TMU0_TUNI0) {
-            timer_clear(TMU0);
-        }
-        else if(evt == EXC_TMU1_TUNI1) {
-            timer_clear(TMU1);
-        }
-        else {
-            timer_clear(TMU2);
-        }
-
-        handled = 1;
-    }
-
     /* If there's a handler, call it */
     {
         hnd = &irq_handlers[evt >> 5];
@@ -355,7 +332,7 @@ void irq_create_context(irq_context_t *context, uint32_t stkpntr,
 static void irq_def_timer(irq_t src, irq_context_t *context, void *data) {
     (void)src;
     (void)context;
-    (void)data;
+    timer_clear((int)data);
 }
 
 /* Default FPU exception handler (can't seem to turn these off) */
@@ -385,15 +362,17 @@ int irq_init(void) {
     irq_disable();
 
     /* Blank the exception handler tables */
-    memset(irq_handlers, 0, sizeof(irq_handlers));
-    memset(trapa_handlers, 0, sizeof(trapa_handlers));
+    memset(irq_handlers,        0, sizeof(irq_handlers));
+    memset(trapa_handlers,      0, sizeof(trapa_handlers));
     memset(&global_irq_handler, 0, sizeof(global_irq_handler));
 
     /* Default to not in an interrupt */
     inside_int = 0;
 
-    /* Set a default timer handler */
-    irq_set_handler(EXC_TMU0_TUNI0, irq_def_timer, (void *)0);
+    /* Set a default timer handlers */
+    irq_set_handler(EXC_TMU0_TUNI0, irq_def_timer, (void *)TMU0);
+    irq_set_handler(EXC_TMU1_TUNI1, irq_def_timer, (void *)TMU1);
+    irq_set_handler(EXC_TMU2_TUNI2, irq_def_timer, (void *)TMU2);
 
     /* Set a trapa handler */
     irq_set_handler(EXC_TRAPA, irq_handle_trapa, trapa_handlers);