+++ doc/CHANGELOG.md
@@ -15,6 +15,7 @@ Platform-specific changes are prefixed with the platform name, otherwise the cha
 - **Dreamcast**: Add support and update toolchain profiles for Newlib 4.5.0, Binutils 2.43.1, and GDB 15.2 [EF]
 - **Dreamcast**: Make m4-single the default floating-point ABI [EF]
 - **Dreamcast**: Add DMAC API [PC]
+- **Dreamcast**: Add basic MMU functionality [PC]
 
 ## KallistiOS version 2.1.1
 - Added pvrtex utility by TapamN to utils [Daniel Fairchild == DF]